removed all static content
