+++ HouseRizz-iOS/Model/HRCartItem.swift
@@ -1,0 +1,13 @@
+//
+//  HRCartItem.swift
+//  HouseRizz-iOS
+//
+//  Created by Krish Mittal on 05/06/24.
+//
+
+import Foundation
+
+struct HRCartItem {
+    let product: HRProduct
+    var quantity: Int
+}
