+++ HouseRizz-iOS/ObservableObject/ViewModel/Buyer/CartViewModel.swift
@@ -8,17 +8,17 @@
 import Foundation
 
 class CartViewModel: ObservableObject {
-    @Published private(set) var products: [CartItem] = []
+    @Published private(set) var products: [HRCartItem] = []
     @Published private(set) var total: Int = 0
 
     func addToCart(product: HRProduct, quantity: Int = 1) {
         if let index = products.firstIndex(where: { $0.product.id == product.id }) {
             products[index].quantity += quantity
         } else {
-            let cartItem = CartItem(product: product, quantity: quantity)
+            let cartItem = HRCartItem(product: product, quantity: quantity)
             products.append(cartItem)
         }
-        total += product.price * quantity
+        total += Int((product.price ?? 0)) * quantity
     }
 
     func removeFromCart(product: HRProduct) {
@@ -29,19 +29,16 @@ class CartViewModel: ObservableObject {
             } else {
                 products.remove(at: index)
             }
-            total -= product.price
+            total -= Int(product.price ?? 0)
         }
     }
 
-    func updateCartItemQuantity(cartItem: CartItem, newQuantity: Int) {
+    func updateCartItemQuantity(cartItem: HRCartItem, newQuantity: Int) {
         if let index = products.firstIndex(where: { $0.product.id == cartItem.product.id }) {
             products[index].quantity = newQuantity
-            total = products.reduce(0) { $0 + ($1.quantity * $1.product.price) }
+            total = products.reduce(0) { $0 + ($1.quantity * Int(($1.product.price ?? 0))) }
         }
     }
 }
 
-struct CartItem {
-    let product: HRProduct
-    var quantity: Int
-}
+
