+++ HouseRizz-iOS/View/Buyer/CategoryPage/ProductCategoryView.swift
@@ -10,6 +10,7 @@ import SwiftUI
 struct ProductCategoryView: View {
     @EnvironmentObject var cartViewModel: CartViewModel
     @State private var selectedProduct: HRProduct?
+    @StateObject private var viewModel = ProductCategoryViewModel()
 
     var column = [GridItem(.adaptive(minimum: 160), spacing: 20)]
     var productCategory: Category
@@ -19,10 +20,23 @@ struct ProductCategoryView: View {
             VStack {
                 SearchView()
                     .padding(.top, 10)
+                
+                /*
+                 LazyVGrid(columns: [GridItem(.flexible()), GridItem(.flexible())], spacing: 16) {
+                     ForEach(viewModel.products.indices, id: \.self) { index in
+                         ProductCardView(product: viewModel.products[index])
+                             .environmentObject(cartViewModel)
+                             .onTapGesture {
+                                 selectedProduct = viewModel.products[index]
+                             }
+                     }
+                 }
+                 .filter { $0.category == productCategory }
+                 */
 
                 ScrollView {
                     LazyVGrid(columns: column) {
-                        ForEach(productList.filter { $0.category == productCategory }, id: \.id) { product in
+                        ForEach(viewModel.products.filter { $0.category == productCategory.title }, id: \.self) { product in
                             ProductCardView(product: product)
                                 .environmentObject(cartViewModel)
                                 .onTapGesture {
